refactor: create folder for sync action components
