+++ src/commands/continue.ts
@@ -7,7 +7,7 @@ import {
   stackOntoBaseRebaseContinuation,
   stackOntoFixContinuation,
 } from '../actions/onto/stack_onto';
-import { repoSyncDeleteMergedBranchesContinuation } from '../actions/sync';
+import { repoSyncDeleteMergedBranchesContinuation } from '../actions/sync/sync';
 import { TMergeConflictCallstack } from '../lib/config/merge_conflict_callstack_config';
 import { TContext } from '../lib/context/context';
 import { PreconditionsFailedError } from '../lib/errors';
